Fixed ubuntu-pre

+++ base-system/ubuntu-pre.sh
@@ -1,19 +1,55 @@
 #!/bin/bash
 
+set -e
+
 # Installing requried packages
 
-sudo apt-get install bison g++ texinfo squashfs-tools gawk make syslinux-utils libgmp-dev libmpfr-dev libmpc-dev
+sudo apt-get install bison g++ texinfo squashfs-tools gawk make syslinux-utils
+
+if [ $(readlink /bin/sh) != "/bin/bash" ]; then
+	ln -svf /bin/bash /bin/sh
+fi
+
+# Installing development libraries
+
+if [ ! -f /usr/lib/libgmp.la ]; then
+	wget http://ftp.gnu.org/gnu/gmp/gmp-6.1.0.tar.xz
+	tar xf gmp-6.1.0.tar.xz 
+	cd gmp-6.1.0/
+	./configure --prefix=/usr && make -j$(nproc) && sudo make install
+	cd ..
+	rm -rf gmp-6.1.0/
+fi
+
+if [ ! -f /usr/lib/libmpfr.la ]; then
+	wget http://www.mpfr.org/mpfr-3.1.3/mpfr-3.1.3.tar.xz
+	tar xf mpfr-3.1.3.tar.xz 
+	cd mpfr-3.1.3/
+	./configure --prefix=/usr && make -j$(nproc) && sudo make install 
+	cd ..
+	rm -rf mpfr-3.1.3/
+fi
+
+if [ ! -f /usr/lib/libmpc.la ]; then
+	wget https://ftp.gnu.org/gnu/mpc/mpc-1.0.3.tar.gz
+	tar xf mpc-1.0.3.tar.gz 
+	cd mpc-1.0.3/
+	./configure --prefix=/usr && make -j$(nproc) && sudo make install 
+	cd ..
+	rm -rf mpc-1.0.3/
+fi
 
 # Installing our version of cdrtools because ubuntu's version is an alias for geniosimage. This is needed to create ISO.
 
-wget https://sourceforge.net/projects/cdrtools/files/cdrtools-3.01.tar.bz2
-tar xf cdrtools-3.01.tar.bz2
-cd cdrtools-3.01/
-make
-cp -v ./mkisofs/OBJ/`uname -m`-linux-cc/mkisofs /usr/bin
-cd ..
-rm -r cdrtools-3.01/
-clear
+if [ ! -e /usr/bin/mkisofs ]; then
+	wget https://sourceforge.net/projects/cdrtools/files/cdrtools-3.01.tar.bz2
+	tar xf cdrtools-3.01.tar.bz2
+	cd cdrtools-3.01/
+	make -j$(nproc)
+	cp -v ./mkisofs/OBJ/`uname -m`-linux-cc/mkisofs /usr/bin
+	cd ..
+	rm -r cdrtools-3.01/
+fi
 
 # Checking once again.
 cat > version-check.sh << "EOF"
